do not spam logs by killing the syslog server repeatedly

+++ pkg/acquisition/modules/syslog/syslog.go
@@ -180,11 +180,15 @@ func (s *SyslogSource) buildLogFromSyslog(ts *time.Time, hostname *string,
 }
 
 func (s *SyslogSource) handleSyslogMsg(out chan types.Event, t *tomb.Tomb, c chan syslogserver.SyslogMessage) error {
+	killed := false
 	for {
 		select {
 		case <-t.Dying():
-			s.logger.Info("Syslog datasource is dying")
-			s.serverTomb.Kill(nil)
+			if !killed {
+				s.logger.Info("Syslog datasource is dying")
+				s.serverTomb.Kill(nil)
+				killed = true
+			}
 		case <-s.serverTomb.Dead():
 			s.logger.Info("Syslog server has exited")
 			return nil
